# Optimizations: symbolic links skip, binary files detection