--- a/.mypy.ini
+++ b/.mypy.ini
@@ -42,3 +42,6 @@ ignore_missing_imports = True
 
 [mypy-setuptools.*]
 ignore_missing_imports = True
+
+[mypy-password_strength.*]
+ignore_missing_imports = True