--- a/credsweeper/app_path.py
+++ b/credsweeper/app_path.py
@@ -1,4 +1,0 @@
-# use separate file to avoid circular import issue
-from pathlib import Path
-""" Directory of credsweeper sources """
-APP_PATH = Path(__file__).resolve().parent