--- a/credsweeper/common/__init__.py
+++ b/credsweeper/common/__init__.py
@@ -1,1 +1,4 @@
 from credsweeper.common.keyword_checklist import KeywordChecklist
+
+# use the variable to avoid singleton creation and testing easiest
+static_keyword_checklist = KeywordChecklist()