--- a/credsweeper/common/__init__.py
+++ b/credsweeper/common/__init__.py
@@ -1,4 +1,4 @@
 from credsweeper.common.keyword_checklist import KeywordChecklist
 
-# use the variable to avoid singleton creation and testing easiest
+# use the variable to avoid singleton creation and make testing easier
 static_keyword_checklist = KeywordChecklist()