--- a/credsweeper/deep_scanner/abstract_scanner.py
+++ b/credsweeper/deep_scanner/abstract_scanner.py
@@ -1,4 +1,4 @@
-from abc import abstractmethod, ABC, abstractproperty
+from abc import abstractmethod, ABC
 from typing import List
 
 from credsweeper.config import Config
@@ -11,12 +11,14 @@ from credsweeper.scanner import Scanner
 class AbstractScanner(ABC):
     """Base abstract class for all recursive scanners"""
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def config(self) -> Config:
         """Abstract property to be defined in DeepScanner"""
         raise NotImplementedError(__name__)
 
-    @abstractproperty
+    @property
+    @abstractmethod
     def scanner(self) -> Scanner:
         """Abstract property to be defined in DeepScanner"""
         raise NotImplementedError(__name__)