--- a/credsweeper/filters/group/__init__.py
+++ b/credsweeper/filters/group/__init__.py
@@ -1,6 +1,6 @@
 from credsweeper.filters.group.group import Group  # isort:skip
 
-from credsweeper.filters.group.credit_card_number_sequense import CreditCardNumberSequence
+from credsweeper.filters.group.credit_card_number_sequence import CreditCardNumberSequence
 from credsweeper.filters.group.general_keyword import GeneralKeyword
 from credsweeper.filters.group.general_pattern import GeneralPattern
 from credsweeper.filters.group.password_keyword import PasswordKeyword