--- a/credsweeper/filters/value_dictionary_keyword_check.py
+++ b/credsweeper/filters/value_dictionary_keyword_check.py
@@ -18,6 +18,8 @@ class ValueDictionaryKeywordCheck(Filter):
         """
         if not line_data.value:
             return True
-        if any(keyword in line_data.value.lower() for keyword in static_keyword_checklist.keyword_set):
-            return True
+        line_data_value_lower = line_data.value.lower()
+        for keyword in static_keyword_checklist.keyword_set:
+            if keyword in line_data_value_lower:
+                return True
         return False