--- a/credsweeper/rules/rule.py
+++ b/credsweeper/rules/rule.py
@@ -51,7 +51,7 @@ class Rule:
         self.pattern_type: str = Rule._get_pattern_type(self.rule_type, len(self.patterns))
         self.use_ml: bool = rule_template["use_ml"]
         self.validations: List[Validation] = rule_template.get("validations")
-        self.required_substrings: List[str] = [s.lower() for s in rule_template.get("required_substrings", [""])]
+        self.required_substrings = [i.strip().lower() for i in rule_template.get("required_substrings", [""])]
         self.min_line_len: int = rule_template.get("min_line_len", -1)
         self.usage_list: List[str] = rule_template.get("usage_list")