… .github/workflows/benchmark.yml
… credsweeper/common/morpheme_checklist.txt
… credsweeper/filters/__init__.py
… credsweeper/filters/group/credit_card_number_sequence.py
… credsweeper/ml_model/features.py
… credsweeper/scanner/scan_type/pem_key_pattern.py
… credsweeper/scanner/scanner.py
… credsweeper/utils/util.py
… requirements.txt
… tests/__init__.py
… tests/common/test_keyword_checklist.py
… tests/config/test_config.py
… tests/conftest.py
… tests/data/depth_9.json
… tests/data/doc.json
… tests/data/output.json
… tests/file_handler/test_file_path_extractor.py
… tests/file_handler/test_patch_provider.py
… tests/filters/test_value_couple_keyword_check.py
… tests/filters/test_value_structured_token_check.py
… tests/filters/test_value_token_base32_check.py
… tests/ml_model/test_features.py
… tests/ml_model/test_ml_validator.py
… tests/samples/auth.asciidoc
… tests/samples/aws_mws_key
… tests/samples/aws_mws_key.txt
… tests/samples/certificate.go
… tests/samples/credential
… tests/samples/credential.txt
… tests/samples/facebook_key